requirements.txt
Cheetah==2.4.4
Django==1.4.1
GnuPGInterface==0.3.2
Landscape-Client==12.05
M2Crypto==0.20.1
MySQL-python==1.2.3
PAM==0.4.2
PIL==1.1.7
PyYAML==3.10
Twisted-Core==11.0.0
Twisted-Names==11.0.0
Twisted-Web==11.0.0
Whoosh==2.4.1
apt-xapian-index==0.44
boto==2.0
chardet==2.0.1
cloud-init==0.6.1
command-not-found==0.2.44
configobj==4.7.2
distribute==0.6.19
django-facebook==4.2.2
django-grappelli==2.4.4
django-haystack==1.2.7
django-registration==0.8
django-rosetta==0.7.1
django-suit==0.1.9
django-tastypie==0.9.14
euca2ools==2.0.0
httplib2==0.7.2
keyring==0.6.2
launchpadlib==1.9.8
lazr.restfulclient==0.11.2
lazr.uri==1.0.2
mimeparse==0.1.3
oauth==1.0.1
paramiko==1.7.7.1
pexpect==2.3
pyOpenSSL==0.12
pycrypto==2.3
pycurl==7.19.0
pyserial==2.5
python-apt==0.8.0ubuntu9
python-dateutil==2.1
python-debian==0.1.20ubuntu2
python-memcached==1.48
simplejson==2.1.6
six==1.3.0
smart==1.4
sorl-thumbnail==11.12
ufw==0.30.1-2ubuntu1
unattended-upgrades==0.1
virtualenv==1.6.4
wadllib==1.2.0
wsgiref==0.1.2
zope.interface==3.6.1
